% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_col_na.R
\name{assess_col_na}
\alias{assess_col_na}
\title{Asssess percentage of missing values in columns}
\usage{
assess_col_na(x3p, perc_of_col = 40, threshold_prop = 0.2)
}
\arguments{
\item{x3p}{scan in x3p format}

\item{perc_of_col}{HH: what is the expected range for the parameter? defaults to 40}

\item{threshold_prop}{defaults to 0.2}
}
\value{
logical value: is percentage of columns with more than XXX percent missing values below XXX?
HH: replace XXX by parameter names
}
\description{
Asssess percentage of missing values in columns
}
\examples{
assess_col_na(fau001_ba_l1)
assess_col_na(fau277_bb_l2)
}
