% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_median_na_proportion.R
\name{assess_median_na_proportion}
\alias{assess_median_na_proportion}
\title{Feature Extraction: Median of the proportion of NA values present in the middle of a 3d scan,
from a specific number of different y values.}
\usage{
assess_median_na_proportion(x3p, chopoff = 1/6, numlines = 20)
}
\arguments{
\item{x3p}{scan in x3p format}

\item{chopoff}{numeric value between 0 and .5 - The percentage of values to chop off of both sides of the scan}

\item{numlines}{Number of horizontal lines across the scan to find the median proportion of na values from}
}
\value{
median proportion of missing values
}
\description{
One of the main problems with 3d scans are the feathering that occurs within the middle of the scan.
Here, we quantify this feathering by taking samples at specific number of y values in the surface matrix
that are in a certain percentage of the middle of the 3d scan. Then with these rows of the surface matrix we measure
the proportion of missing values in each of those rows. We then take the median proportion of missing values
of those rows, which will give us a good measure to quantify the amount of feathering in a 3d scan.
}
\examples{
data(fau277_bb_l2)
assess_median_na_proportion(fau277_bb_l2)
}
