% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_middle_na_proportion.R
\name{assess_middle_na_proportion}
\alias{assess_middle_na_proportion}
\title{Feature Extraction: Proportion of NA values in the middle of the scan.}
\usage{
assess_middle_na_proportion(x3p, chopoff = 1/8)
}
\arguments{
\item{x3p}{scan in x3p format}

\item{chopoff}{numeric value between 0 and .5 - The percentage of values to chop off of both sides of the scan}
}
\value{
proportion of NA values in the middle of 3d scan
}
\description{
One of the main problems with 3d scans is NA values in the middle of the 3d scan.
Here we are quantifying this by chopping off a portion of the right and left sides of a 3d scan
and are then finding the proportion of missing values in the middle of the 3d scan.
}
\examples{
data(fau277_bb_l2)
assess_middle_na_proportion(fau277_bb_l2)
}
