% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_median_na_proportion.R
\name{extract_median_na_proportion}
\alias{extract_median_na_proportion}
\title{Feature Extraction: Median of the proportion of NA values present in the middle of a 3d scan,
from 20 different y values.}
\usage{
extract_median_na_proportion(x3p)
}
\arguments{
\item{x3p}{scan in x3p format}
}
\value{
median percentage of missing values # HH: percentage or proportion?
HH: I believe, that what you are calculating is actually a lot more complicated than an overall median of missing values.
}
\description{
One of the main problems with 3d scans are the feathering that occurs within the middle of the scan.
Here, we quantify this feathering by taking samples at 20 specific y values in the surface matrix
that are in the middle 2/3 of the image. Then with these rows of the surface matrix we measure
the proportion of missing values in each of those 20 rows. We then take the median proportion of missing values
of those rows, which will give us
a good measure to quantify the amount of feathering in a 3d scan.
}
\examples{
data(fau277_bb_l2)
extract_median_na_proportion(fau277_bb_l2)
}
