% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_na.r
\name{extract_na}
\alias{extract_na}
\title{Feature extraction: percentage of missing values}
\usage{
extract_na(x3p)
}
\arguments{
\item{x3p}{scan in x3p format}
}
\value{
percentage of missing values in the surface matrix of the scan
}
\description{
A good quality of 3d scans is crucial for an assessment of similarity of
striation marks between different bullets.
Here, we measure the percentage of missing values in the scan. Generally,
a higher percentage of missing values is associated with a lower scan quality.
}
\examples{
data(fau277_bb_l2)
extract_na(fau277_bb_l2) # this scan has a particularly high percentage of missing values
}
