% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_na_column.r
\name{extract_na_column}
\alias{extract_na_column}
\title{Feature extraction: percentage of missing values by column}
\usage{
extract_na_column(x3p)
}
\arguments{
\item{x3p}{scan in x3p format}
}
\value{
percentage of missing values in each column of the scan's surface matrix
}
\description{
A good quality of 3d scans is crucial for an assessment of similarity of
striation marks between different bullets.
Here, we measure the percentage of missing values in the scan. Generally,
a higher percentage of missing values is associated with a lower scan quality.
}
\examples{
data(fau277_bb_l2)
nas_bad <- extract_na_column(fau277_bb_l2)
plot(nas_bad) # the 'feathering' becomes obvious in the spikes of the missing values

nas_good <- extract_na_column(fau001_ba_l1) # good scan
plot((1:length(nas_good))*3, nas_good, col =2)
points(nas_bad, col=4)
}
