% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{fau001_ba_l1}
\alias{fau001_ba_l1}
\title{Example x3p scan}
\format{
An object of class \code{x3p} of length 5.
}
\usage{
fau001_ba_l1
}
\description{
This is an x3p file containing an example of an excellent scan (sub-sampled)
}
\examples{
data(fau001_ba_l1)
x3p_image(fau001_ba_l1)
}
\keyword{datasets}
