% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{fau277_bb_l2}
\alias{fau277_bb_l2}
\title{Example x3p scan}
\format{
An object of class \code{x3p} of length 5.
}
\usage{
fau277_bb_l2
}
\description{
This is an x3p file containing an example of a very bad scan
}
\keyword{datasets}
