% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_brightness.R
\name{find_brightness}
\alias{find_brightness}
\title{Get the lighting from an x3p object}
\usage{
find_brightness(x3p)
}
\arguments{
\item{x3p}{is a scan in x3p format}
}
\value{
the brightness of the file from the Comment attribute of the x3p object
}
\description{
Get the lighting from an x3p object
}
