% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_brightness.R
\name{lighting_protocol}
\alias{lighting_protocol}
\title{Get lighting protocol from the name of an x3p object}
\usage{
lighting_protocol(name)
}
\arguments{
\item{name}{name of an x3p object}
}
\value{
1 for 20\% autolight, and 2 for 20\% x 10 autolight ('flooding')
}
\description{
Get lighting protocol from the name of an x3p object
}
