% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_quality.R
\name{predict_quality}
\alias{predict_quality}
\title{Predicts the quality of a vector of x3p scan's by outputing the random forest predicted probability that a scan's quality is good.}
\usage{
predict_quality(x3pvector)
}
\arguments{
\item{x3pvector}{a vector of x3p objects to be classified as good or bad quality scans}
}
\value{
a vector of random forest probability scores
}
\description{
Predicts the quality of a vector of x3p scan's by outputing the random forest predicted probability that a scan's quality is good.
}
