% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_quality.R
\name{predict_quality_one}
\alias{predict_quality_one}
\title{Predicts the probability that a x3p scan will be good and also predicts the reason why it is bad quality if it is predicted to be bad quality.}
\usage{
predict_quality_one(x3p, cutoff = 0.57)
}
\arguments{
\item{x3p}{x3p scan of a bullet land}

\item{cutoff}{numeric value in \link{0,1} used as cut off between bad scans and good ones.}
}
\value{
a dataframe with one row, containing all feature values and a prediction of the scan quality
}
\description{
Predicts the probability that a x3p scan will be good and also predicts the reason why it is bad quality if it is predicted to be bad quality.
}
\examples{
data(fau277_bb_l2)
predict_quality_one(fau277_bb_l2)
}
