% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{randomforest}
\alias{randomforest}
\title{Randomforest predicting oveeall scan quality}
\format{
An object of class \code{randomForest.formula} (inherits from \code{randomForest}) of length 19.
}
\usage{
randomforest
}
\description{
some more text
}
\examples{
data(randomforest)
randomforest
}
\keyword{datasets}
