% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{randomforest2}
\alias{randomforest2}
\title{Randomforest predicting reason for lack of quality}
\format{
An object of class \code{randomForest.formula} (inherits from \code{randomForest}) of length 19.
}
\usage{
randomforest2
}
\description{
some more text
}
\examples{
data(randomforest2)
randomforest2
}
\keyword{datasets}
