% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_quality.R
\name{predict_quality}
\alias{predict_quality}
\title{Predicts the probability that a x3p scan will be good and also predicts the reason why it is bad quality if it is predicted to be bad quality.}
\usage{
predict_quality(x3pobjectvector, x3pnamevector)
}
\arguments{
\item{x3pobjectvector}{a vector of x3p objects to be classified as good or bad quality scans}

\item{x3pnamevector}{a vector of names of x3p files}
}
\value{
a dataframe containing a vector of predicted probabilities of a scan being good and its predicted quality_type
}
\description{
Predicts the probability that a x3p scan will be good and also predicts the reason why it is bad quality if it is predicted to be bad quality.
}
